refactor: refactor sign up logic
