+++ src/main/java/com/example/comitserver/dto/UserResponseDTO.java
@@ -5,13 +5,13 @@ import lombok.Data;
 @Data
 public class UserResponseDTO {
     private Long id;
-    private String username;
+    private String name;
+    private String fullName;
     private String phoneNumber;
     private String studentId;
     private String email;
-    private String position;
-    private String bio;
+    private String imageSrc;
     private String github;
     private String blog;
-    private String profileImage;
+
 }
